Refactor connect helpers into dedicated cli_connect module
